Partly revert archiving, split archives per site
A single JSON file containing all the historical archives of tallies for all sites is going to grow to an unmanageable size. Instead, split the historical archives into wiki-specific JSON files, but still keep a general tally with the latest tally for each wiki. This creates some duplication of data, but is easier than both alternatives (having everything in one big file, or having everything only in hundreds of files that we need to go through to display the overview on the main page).

Archives for global tally and family tallies aren't handled yet.