Add script to run periodic tests
Because we're running the main script on hundreds of wikis, it's difficult to identify bugs in the code unless we check every dashboard individually. This test script (planned to be run daily) performs check for every wiki to uncover bugs in a timely manner.

For the moment, we're only checking two things:
* whether the tallies for each wiki have been updated in the last three days
* whether the number of total files is greater than the number of files missing machine-readable metadata (which seems impossible, but is actually happening now for ukwiki).

More tests may be added in the future as appropriate.